adaptive-ir-system/requirements.txt
# Python
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Information Retrieval
pyserini>=0.21.0
rank-bm25>=0.2.2
pytrec-eval>=0.5

# Data Processing
h5py>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8
scikit-learn>=1.2.0

# ML Utilities
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
colorlog>=6.7.0
